Initial working verions for all sources
