+++ radiospectra/spectrogram/__init__.py
@@ -1,1 +1,7 @@
-from radiospectra.spectrogram.spectrogram import Spectrogram, SWAVESSpectrogram
+from radiospectra.spectrogram.sources import (
+    CALISTOSpectrogram,
+    EOVSASpectrogram,
+    RFSSpectrogram,
+    SWAVESSpectrogram,
+)
+from radiospectra.spectrogram.spectrogram import Spectrogram
